+++ tests/api/StoreApiCest.php
@@ -82,6 +82,48 @@ class StoreApiCest
 		$I->dontSeeResponseContainsJson(['phone' => $this->store['telefon']]);
 	}
 
+	public function patchStoreTeamStatusOk(ApiTester $I)
+	{
+		$I->login($this->manager[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 2]);
+		$I->seeResponseCodeIs(Http::OK);
+		$I->seeResponseIsJson();
+		$I->seeResponseContainsJson(['id' => $this->store[self::ID]]);
+		$I->seeResponseContainsJson(['phone' => $this->store['telefon']]);
+	}
+
+	public function patchStoreTeamStatusNotFound(ApiTester $I)
+	{
+		$I->login($this->manager[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID] + 1, ['teamStatus' => 2]);
+		$I->seeResponseCodeIs(Http::NOT_FOUND);
+	}
+
+	public function patchStoreTeamStatusInvalid(ApiTester $I)
+	{
+		$I->login($this->manager[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 3]);
+		$I->seeResponseCodeIs(Http::BAD_REQUEST);
+
+		$I->login($this->manager[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 'a']);
+		$I->seeResponseCodeIs(Http::BAD_REQUEST);
+	}
+
+	public function canPatchOnlyAccessStoreAsFoodsaver(ApiTester $I)
+	{
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 2]);
+		$I->seeResponseCodeIs(Http::UNAUTHORIZED);
+
+		$I->login($this->foodsharer[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 2]);
+		$I->seeResponseCodeIs(Http::FORBIDDEN);
+
+		$I->login($this->teamMember[self::EMAIL]);
+		$I->sendPATCH(self::API_STORES . '/' . $this->store[self::ID], ['teamStatus' => 2]);
+		$I->seeResponseCodeIs(Http::FORBIDDEN);
+	}
+
 	public function canWriteStoreWallpostAndGetAllPosts(ApiTester $I): void
 	{
 		$I->login($this->teamMember[self::EMAIL]);