+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - At stores under pickup history, storemanagers can now see the signed out foodsavers from pickup slots from past 7 days with the signout date. !2363 @martincodes-de
 - Voter Turnout Percentage can be seen if the poll has completed !2377 @viveknshah
 - Provide RestAPI to get backend defined common meta information for stores like "max count of slots per pickup" !1373 @koemai
+- Provide RestAPI to change teamstatus of a store !2406
 
 ## Changes
 - Checks in the id card creation function if only one foodsaver is selected and then reduces the